src/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
openai==0.27.8
python-dotenv==1.0.0
prometheus-client==0.17.1
